updates to latest duneanaobj standalone
- disables NDGAr for the moment as it isn't set up to use the latest duneanaobj

+++ samplePDFDUNE/StructsDUNE.h
@@ -56,7 +56,7 @@ struct dunemc_base {
   double *beam_w;
   double *flux_w;
 
-  int *mode;
+  double *mode;
   int *isbound;
 
   double *rw_truecz;
